696 active tag add requirement fix (#975)
Adding a requirement to a active tag was causing an exception. The requirement would be added, and a preservation period would also be added for the tag requirement.
This resulted in an event being created for the preservation period entity creation. When the creation event was being handled the logic retrieved the parent tag requirement. Since the tag requirement had not been stored yet, this resulted in an exception.

Fixed by refactoring the logic to send a child added event from the tag requirement, that includes information about the parent tag requirement and the child preservation period. This circumvents the need to retrieve the parent when handling the event.

Covering this use case by added an integration test for adding tag requirements to a active tag.

equinor/cc-toolbox#696