824 Field events fixes (#987)
* Fixing an issue where updating fields would cause an error

The "GetRequirementDefinitionByFieldGuidAsync" in the RequirementTypeRepository was not working as intended, and only returned null values.
Fixed the issue by bypassing the need to retrieve the parent object and instead making use of the child modified event logic flow.

* Fixing an issue where sending deletion events would silently fail

Events were being sent to the publisher as a interface, IDeleteEventV1. Mass transit was unable to handle these events, and they were silently ignored.
Changing the implementation to pass the objects directly, instead of a interface representation.